Создан первый класс по Фабрике Вопросов
